Add JavaDoc comments to sanbot.command package
