+++ docs/README.md
@@ -1,4 +1,4 @@
-# sanfobot.Sanbot User Guide
+# sanbot.Sanbot User Guide
 
 // Update the title above to match the actual product name
