+++ src/main/java/sanbot/task/TodoTask.java
@@ -1,4 +1,4 @@
-package sanfobot.task;
+package sanbot.task;
 
 public class TodoTask extends Task {
     public TodoTask(String name) {
