--- src/main/java/sanfobot/command/Command.java
@@ -1,9 +1,0 @@
-package sanfobot.command;
-
-import sanfobot.task.Task;
-
-import java.util.ArrayList;
-
-public interface Command {
-    public void execute(String input, ArrayList<Task> taskList);
-}
