--- src/main/java/sanfobot/command/MarkNotDoneCommand.java
@@ -1,21 +1,0 @@
-package sanfobot.command;
-
-import sanfobot.task.Task;
-import sanfobot.ui.Ui;
-
-import java.util.ArrayList;
-
-public class MarkNotDoneCommand {
-    public void execute(String input, ArrayList<Task> taskList) {
-        String[] processed = input.split(" ");
-        try {
-            int num = Integer.parseInt(processed[1].trim()) - 1;
-            taskList.get(num).markAsNotDone();
-            Ui.markNotDone(taskList.get(num));
-        } catch (ArrayIndexOutOfBoundsException e) {
-            Ui.markNotDoneIndexError();
-        } catch (IndexOutOfBoundsException e) {
-            Ui.markNotDoneInvalidIndex();
-        }
-    }
-}
