+++ src/test/java/sanbot/parser/ParserTest.java
@@ -1,4 +1,4 @@
-package sanfobot.parser;
+package sanbot.parser;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertFalse;
