+++ src/test/java/sanbot/task/EventTaskTest.java
@@ -1,7 +1,7 @@
-package sanfobot.task;
+package sanbot.task;
 
 import org.junit.jupiter.api.Test;
-import sanfobot.command.AddEventTaskCommand;
+import sanbot.command.AddEventTaskCommand;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.ArrayList;
